Trunkline/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.0.1
matplotlib>=3.4.2
seaborn>=0.11.2

# Additional ML libraries
shap>=0.40.0
tqdm>=4.62.0

# Jupyter for tutorials and exploration
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.5

# Utilities
PyYAML>=6.0
argparse>=1.4.0
